[#143] feat(card) - 카드 삭제시 session.username과 다른 카드는 제거 불가

+++ public/javascripts/components/Column.js
@@ -244,6 +244,9 @@ export default class Column {
       (card) => card.getId() === removedCardId
     )
 
+    const username = localStorage.getItem('username')
+    if (username !== removedCard.getUsername()) return
+
     const modal = new DeleteCardModal(removedCard.getTitle(), () => {
       this.removeCard(removedCard)
       removedCard.removeTarget()
